% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{ComputeKs}
\alias{ComputeKs}
\title{A U-statistics method for association test using multi-layer omics data.}
\usage{
ComputeKs(Omics, Y, Xcov = NULL, Kernel, c = 0, d = 2,
  impute = TRUE)
}
\arguments{
\item{Omics}{A list with each element being a n by p matrix. If there is missing, then should be NA for the subject.}

\item{Y}{The continuous or discrete outcomes for $n$ subjects.}

\item{Xcov}{A matrix of demographic variables (e.g., age and gender). The default is NULL.}

\item{Kernel}{The method to calculate the pair-wise sample similarity matrix. It takes the method \code{AM}, \code{IBS}, \code{Lin0}, 
\code{Quad1},\code{Minkowski}, \code{Polyk}.Details can be found in \code{varComp}package.}

\item{impute}{Whether to impute the missing values with the mean. Default is \code{TRUE}.}
}
\value{
A list with the outcome vector \code{Y}, the demographic variable \code{X}, and the pair-wise similarities \code{Ks}
}
\description{
A U-statistics method for association test using multi-layer omics data.
}
\examples{
data(Omics)
data(outcome)
data(Cov)
Input<-ComputeKs(Omics,Y=outcome$all,Xcov=NULL,Kernel=c("Lin0","Lin0","IBS"))
Input<-ComputeKs(Omics,Y=outcome$all,Xcov=Cov,Kernel=c("Lin0","Lin0","IBS"))

}
