% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCovariates.R
\docType{data}
\name{Cov}
\alias{Cov}
\title{Simulated covaraites.}
\format{A data frame with 500 rows and 2 columns:
\describe{
  \item{X1}{Continuous covariate.}
  \item{X2}{Binary covariate.}
  
}}
\usage{
Cov
}
\description{
Simulated covaraites.
}
\keyword{datasets}
