% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wU.R
\name{wU}
\alias{wU}
\title{A U-statistics method for association test using multi-layer omics data.}
\usage{
wU(Y, Ks, X = NULL, scheme = 0, B = NULL)
}
\arguments{
\item{Y}{The continuous or discrete outcomes of subjects.}

\item{Ks}{A list with each element being a pair-wise similarity matrix derived from a layer of omcis data.}

\item{X}{A matrix of demographic variables (e.g., age and gender).}

\item{scheme}{Specify the weighting schemes (0: average weighting, 1: consensus weighting, 2: permutation-based weighting)}

\item{B}{The number of permutations (this is ignored if \code{scheme} is not equal 2). It can be set as null, and the algorithm will determine B.}
}
\value{
The P-value of our proposed method.
}
\description{
A U-statistics method for association test using multi-layer omics data.
}
\examples{
data(Omics)
data(outcome)
data(Cov)
Input<-ComputeKs(Omics,Y=outcome$all,Xcov=Cov,Kernel=c("Lin0","Lin0","IBS"))
p_ave<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=0)
p_con<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=1)
p_perm<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=2,B=1000)
Input<-ComputeKs(Omics,Y=outcome$expr,Xcov=Cov,Kernel=c("Lin0","Lin0","IBS"))
p_ave<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=0)
p_con<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=1)
p_perm<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=2,B=1000)
Input<-ComputeKs(Omics,Y=outcome$null,Xcov=Cov,Kernel=c("Lin0","Lin0","IBS"))
p_ave<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=0)
p_con<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=1)
p_perm<-wU(Y=Input$Y,Ks=Input$Ks,X=Input$X,scheme=2)
}
